+++ resources/views/components/packed-bubble.blade.php
@@ -1,0 +1,5 @@
+@props(['nodes' => [], 'categories' => []])
+
+<x-chart::base component="packed-bubble" :option="\WireCharts\Support\Option::packedBubble([
+    'nodes' => $nodes, 'categories' => $categories,
+])" {{ $attributes }} />