+++ resources/views/components/scatter-symbols.blade.php
@@ -1,0 +1,5 @@
+@props(['series' => [], 'legend' => true])
+
+<x-chart::base component="scatter-symbols" :option="\WireCharts\Support\Option::scatterSymbols([
+    'series' => $series, 'legend' => $legend,
+])" {{ $attributes }} />